--- requirements.txt.txt
@@ -1,10 +1,0 @@
-streamlit>=1.31.0
-groq>=0.3.2
-python-dotenv>=1.0.0
-dataclasses>=0.6
-typing>=3.7.4
-pathlib>=1.0.1
-logging>=0.5.1.2
-python-dateutil>=2.8.2
-watchdog>=3.0.0
-protobuf>=4.25.1
